update install scripts

+++ git/install.sh
@@ -1,7 +1,13 @@
 #!/bin/sh
 
+echo -e "git configuration...\n"
+
 # backup existing ~/.gitconfig file
-mv ~/.gitconfig ~/.gitconfig.orig
+if [ -f ~/.gitconfig ]; then
+	mv ~/.gitconfig ~/.gitconfig.orig
+fi
 
 # create symbolic link to the gitconfig
 ln -s $DOTFILES/git/gitconfig ~/.gitconfig
+
+echo -e "\ndone...\n"
