+++ vim/install.sh
@@ -1,12 +1,16 @@
 #!/bin/sh
 
+echo -e "vim configuraton...\n"
+
 # install vundle
 if [ ! -d ~/.vim/bundle/vundle/ ]; then
 	git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
 fi
 
 # backup existing ~/.vimrc file
-cp ~/.vimrc ~/.vimrc.orig
+if [ -f ~/vimrc ]; then
+	cp ~/.vimrc ~/.vimrc.orig
+fi
 
 # overwrite vimrc
 cat > ~/.vimrc << "EOF"
@@ -14,3 +18,5 @@ if filereadable(expand("/home/ilias/.dotfiles/vim/vimrc"))
     source /home/ilias/.dotfiles/vim/vimrc
 endif
 EOF
+
+echo -e "\ndone...\n"
